+++ packages/integration-sdk-cli/src/commands/visualize.ts
@@ -1,18 +1,26 @@
 import { createCommand } from 'commander';
+import path from 'path';
 
 import * as log from '../log';
 import { generateVisualization } from '../visualization';
 
 export function visualize() {
   return createCommand('visualize')
-    .description('Visualizes the collected data as a graph')
+    .description(
+      'generate graph visualization of collected entities and relationships',
+    )
     .option(
-      '--data-dir <relative_directory>',
-      'The directory used to generate Nodes/Edges for visualization',
-      '.j1-integration/graph',
+      '-d, --data-dir <directory>',
+      'absolute path to collected entities and relationships',
+      path.resolve(process.cwd(), '.j1-integration', 'graph'),
+    )
+    .option(
+      '-o, --output-file <path>',
+      'absolute path of generated HTML file',
+      path.resolve(process.cwd(), '.j1-integration', 'index.html'),
     )
     .action(async (options) => {
-      const graphLocation = await generateVisualization(options.dataDir);
-      log.info(`Visualize graph here: ${graphLocation}`);
+      await generateVisualization(options.dataDir, options.outputFile);
+      log.info(`Visualize graph here: ${options.outputFile}`);
     });
 }